requirements.txt
asgiref==3.5.1
astroid==2.11.5
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
behave==1.2.6
behave-django==1.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coverage==6.3.2
cryptography==37.0.2
dill==0.3.4
Django==4.0.4
factory-boy==3.2.1
Faker==13.11.1
h11==0.13.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
outcome==1.1.0
parse==1.19.0
parse-type==0.6.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
pylint==2.13.8
pylint-plugin-utils==0.7
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.27.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
wrapt==1.14.1
wsproto==1.1.0
